chore: move utils method
