+++ webapp/src/modules/nft/estate/utils.ts
@@ -1,7 +1,8 @@
 import { NFTCategory } from '@dcl/schemas'
 import { getConnectedProvider } from 'decentraland-dapps/dist/lib/eth'
 import { EstateRegistry__factory } from '../../../contracts'
-import { getContract, getCurrentSigner } from '../../contract/utils'
+import { getCurrentSigner } from '../../wallet/utils'
+import { getContract } from '../../contract/utils'
 import { NFT } from '../types'
 
 export const getSelection = (estate: NFT['data']['estate']) => {
