+++ webapp/src/modules/wallet/utils.ts
@@ -27,3 +27,13 @@ export async function getEth(): Promise<ethers.providers.Web3Provider> {
 
   return new ethers.providers.Web3Provider(provider)
 }
+
+export async function getCurrentSigner() {
+  const provider = await getConnectedProvider()
+  if (!provider) {
+    throw new Error('Could not connect to provider')
+  }
+
+  const eth = new ethers.providers.Web3Provider(provider)
+  return eth.getSigner()
+}
